requirements.txt
numpy
pandas
transformers
torch
pillow
tqdm
streamlit
